feat(cache): implement caching mechanism with local and synchronized caches

+++ identica-api/src/main/java/me/whereareiam/identica/model/config/Settings.java
@@ -70,7 +70,7 @@ public class Settings {
 			private String password;
 
 			private boolean ssl;
-			private long timeout;
+			private int timeout;
 
 			private Channels channels;
 		}
